// web/gen/api.ts
export interface PrismResponse {
    /**
     * @generated from protobuf field: uint64 prism_balance = 1
     */
    prismBalance: bigint;
    /**
     * @generated from protobuf field: uint64 prism_unredeemed = 2
     */
    prismUnredeemed: bigint;
    /**
     * @generated from protobuf field: uint64 prism_redeemable = 3
     */
    prismRedeemable: bigint;
}
/**
 * @generated from protobuf message api.LOMreward
 */
export interface LOMreward {
    /**
     * @generated from protobuf field: string market_id = 1
     */
    marketId: string;
    /**
     * @generated from protobuf field: string account_id = 2
     */
    accountId: string;
    /**
     * @generated from protobuf field: float distance = 3
     */
    distance: number;
    /**
     * @generated from protobuf field: float size = 4
     */
    size: number;
    /**
     * @generated from protobuf field: float duration = 5
     */
    duration: number;
    /**
     * @generated from protobuf field: float lom_score = 6
     */
    lomScore: number;
    /**
     * @generated from protobuf field: string created_at = 7
     */
    createdAt: string;
}
// ...
class PrismResponse$Type extends MessageType<PrismResponse> {
    constructor() {
        super("api.PrismResponse", [
            { no: 1, name: "prism_balance", kind: "scalar", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 2, name: "prism_unredeemed", kind: "scalar", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 3, name: "prism_redeemable", kind: "scalar", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ }
        ]);
    }
    create(value?: PartialMessage<PrismResponse>): PrismResponse {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.prismBalance = 0n;
        message.prismUnredeemed = 0n;
        message.prismRedeemable = 0n;
        if (value !== undefined)
            reflectionMergePartial<PrismResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: PrismResponse): PrismResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* uint64 prism_balance */ 1:
                    message.prismBalance = reader.uint64().toBigInt();
                    break;
                case /* uint64 prism_unredeemed */ 2:
                    message.prismUnredeemed = reader.uint64().toBigInt();
                    break;
                case /* uint64 prism_redeemable */ 3:
                    message.prismRedeemable = reader.uint64().toBigInt();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: PrismResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* uint64 prism_balance = 1; */
        if (message.prismBalance !== 0n)
            writer.tag(1, WireType.Varint).uint64(message.prismBalance);
        /* uint64 prism_unredeemed = 2; */
        if (message.prismUnredeemed !== 0n)
            writer.tag(2, WireType.Varint).uint64(message.prismUnredeemed);
        /* uint64 prism_redeemable = 3; */
        if (message.prismRedeemable !== 0n)
            writer.tag(3, WireType.Varint).uint64(message.prismRedeemable);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message api.PrismResponse
 */
export const PrismResponse = new PrismResponse$Type();
// @generated message type with reflection information, may provide speed optimized methods
class LOMreward$Type extends MessageType<LOMreward> {
    constructor() {
        super("api.LOMreward", [
            { no: 1, name: "market_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "account_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "distance", kind: "scalar", T: 2 /*ScalarType.FLOAT*/ },
            { no: 4, name: "size", kind: "scalar", T: 2 /*ScalarType.FLOAT*/ },
            { no: 5, name: "duration", kind: "scalar", T: 2 /*ScalarType.FLOAT*/ },
            { no: 6, name: "lom_score", kind: "scalar", T: 2 /*ScalarType.FLOAT*/ },
            { no: 7, name: "created_at", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<LOMreward>): LOMreward {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.marketId = "";
        message.accountId = "";
        message.distance = 0;
        message.size = 0;
        message.duration = 0;
        message.lomScore = 0;
        message.createdAt = "";
        if (value !== undefined)
            reflectionMergePartial<LOMreward>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: LOMreward): LOMreward {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string market_id */ 1:
                    message.marketId = reader.string();
                    break;
                case /* string account_id */ 2:
                    message.accountId = reader.string();
                    break;
                case /* float distance */ 3:
                    message.distance = reader.float();
                    break;
                case /* float size */ 4:
                    message.size = reader.float();
                    break;
                case /* float duration */ 5:
                    message.duration = reader.float();
                    break;
                case /* float lom_score */ 6:
                    message.lomScore = reader.float();
                    break;
                case /* string created_at */ 7:
                    message.createdAt = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: LOMreward, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string market_id = 1; */
        if (message.marketId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.marketId);
        /* string account_id = 2; */
        if (message.accountId !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.accountId);
        /* float distance = 3; */
        if (message.distance !== 0)
            writer.tag(3, WireType.Bit32).float(message.distance);
        /* float size = 4; */
        if (message.size !== 0)
            writer.tag(4, WireType.Bit32).float(message.size);
        /* float duration = 5; */
        if (message.duration !== 0)
            writer.tag(5, WireType.Bit32).float(message.duration);
        /* float lom_score = 6; */
        if (message.lomScore !== 0)
            writer.tag(6, WireType.Bit32).float(message.lomScore);
        /* string created_at = 7; */
        if (message.createdAt !== "")
            writer.tag(7, WireType.LengthDelimited).string(message.createdAt);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message api.LOMreward
 */
export const LOMreward = new LOMreward$Type();